+++ tests/rx/rx_mock.hpp
@@ -10,6 +10,8 @@ struct RxMock : dcc::rx::CrtpBase<RxMock> {
   MOCK_METHOD(void, function, (uint16_t, uint32_t, uint32_t), ());
   MOCK_METHOD(void, serviceModeHook, (bool), ());
   MOCK_METHOD(void, serviceAck, (), ());
+  MOCK_METHOD(void, transmitBiDi, (std::span<uint8_t const>), ());
+  MOCK_METHOD(void, error, (), ());
   MOCK_METHOD(uint8_t, readCv, (uint32_t), ());
   MOCK_METHOD(uint8_t, readCv, (uint32_t, uint8_t), ());
   MOCK_METHOD(uint8_t, writeCv, (uint32_t, uint8_t), ());
@@ -23,6 +25,5 @@ struct RxMock : dcc::rx::CrtpBase<RxMock> {
               writeCv,
               (uint32_t, uint8_t, std::function<void(uint8_t)>),
               ());
-  MOCK_METHOD(void, transmitBiDi, (std::span<uint8_t const>), ());
   MOCK_METHOD(void, eastWestDirection, (uint16_t, std::optional<int32_t>), ());
 };